Extract resourceUtils
