+++ CliClient/tests/enex_to_html/en-media-audio.html
@@ -4,7 +4,7 @@
       <source src=":/9168ee833d03c5ea7c730ac6673978c1" type="audio/mp4">
       <p>Your browser does not support HTML5 audio.</p>
     </audio>
-    <p><strong>Download Audio:</strong><a onclick="ipcProxySendToHost(" joplin:="" 9168ee833d03c5ea7c730ac6673978c1");="" return="" false;"="" href=":/9168ee833d03c5ea7c730ac6673978c1">M4A</a>,</p>
+    <p><strong>Download Audio:</strong><a onclick="ipcProxySendToHost('joplin://9168ee833d03c5ea7c730ac6673978c1'); return false;" href=":/9168ee833d03c5ea7c730ac6673978c1">M4A</a></p>
   </div>
   <div>
     <br>
